Refactor: css, js
